requirements.txt
flask==1.1.1
keras==2.2.4
tensorflow==1.14.0
Pillow==6.1.0
requests==2.22.0
gunicorn==19.9.0
